+++ hooks/queries/my/pets/usePostPetsQueries.ts
@@ -1,11 +1,9 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import PetsApi, { PetsAddDataType } from '@/api/my/pets';
-import { useRouter } from 'next/router';
 
 const petsApi = new PetsApi();
 
 export const usePetsModifyQuery = () => {
-  const router = useRouter();
   const queryClient = useQueryClient();
 
   return useMutation({
@@ -15,13 +13,11 @@ export const usePetsModifyQuery = () => {
     },
     onSuccess: (data, variables) => {
       queryClient.invalidateQueries({ queryKey: ['petsDetail', variables.petsId], refetchType: 'active' });
-      router.push('/start-pet/finish');
     },
   });
 };
 
 export const usePetsAddQuery = () => {
-  const router = useRouter();
   const queryClient = useQueryClient();
 
   return useMutation({
@@ -31,7 +27,6 @@ export const usePetsAddQuery = () => {
     },
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['pets'], refetchType: 'active' });
-      router.push('/start-pet/finish');
     },
   });
 };
