+++ pages/mypage/pet/[id]/index.tsx
@@ -1,0 +1,5 @@
+import ModifyPetForm from '@/components/common/modify-pet-form/modify-pet-form';
+
+export default function MypagePetModify() {
+  return <ModifyPetForm routeUrl="/mypage/pet" />;
+}
