+++ pages/mypage/pet/add/index.tsx
@@ -1,0 +1,5 @@
+import AddPetForm from '@/components/common/add-pet-form/add-pet-form';
+
+export default function PetAddPage() {
+  return <AddPetForm routeUrl="/mypage/pet" />;
+}
