--- pages/start-pet/index.module.scss
@@ -1,3 +1,0 @@
-.section {
-  padding: 126px 24px 51px;
-}
